requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
gunicorn==21.2.0
packaging==23.1
paho-mqtt==1.6.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
websockets==11.0.3
